# src/main/resources/static/api-docs.yaml
openapi: 3.0.1
info:
  title: Foro Alura API
  description: |-
    Practice project - Alura Challenge ONE
    para más información visitar el repositorio [Foro Alura](https://github.com/SaulMMBP/foro-alura)
  contact:
    name: Saul Malagon Martinez
    email: smalagonmtz@gmail.com
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
  /usuarios:
    get:
      tags:
      - Usuarios
      summary: Obtiene una lista paginada de todos los usuarios
      operationId: getUsuarios
      parameters:
      - name: page
        in: query
        description: Índice de página basada en 0 (0..N)
        required: false
        schema:
          minimum: 0
          type: integer
          default: 0
      - name: size
        in: query
        description: El tamaño de la página a devolver.
        required: false
        schema:
          minimum: 1
          type: integer
          default: 10
      - name: sort
        in: query
        description: "Criterio de ordenamiento en formato: `property,(asc|desc)`. Default\
          \ sort order is ascending (asc). Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/PagedUsuarioResponse'
    put:
      tags:
      - Usuarios
      summary: Modifica la información de un usuario
      description: "Solo los usuarios con rol `ADMIN` pueden modificar la informació\
        n de cualquier usuario, los usuarios sin rol `ADMIN` solo pueden modificar\
        \ su propia información."
      operationId: updateUsuario
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioPutRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not Found
        "403":
          description: Forbidden
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/UsuarioResponse'
      security:
      - Foro Alura Auth: []
    post:
      tags:
      - Usuarios
      summary: Crea un nuevo usuario
      description: Se debe usar una contraseña encriptada con bcrypt para registrar
        un nuevo usuario
      operationId: newUsuario
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioPostRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Forbidden
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/UsuarioResponse'
...
  /usuarios/{id}:
    get:
      tags:
      - Usuarios
      summary: Obtiene un usuario por su id
      operationId: getUsuario
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/UsuarioResponse'
    delete:
      tags:
      - Usuarios
      summary: Elimina un usuario por su id
      description: "Solo los usuarios con rol `ADMIN` puden eliminar cualquier usuario,\
        \ los usuarios sin rol `ADMIN` solo pueden eliminar su propia cuenta."
      operationId: deleteUsuario
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "403":
          description: Forbidden
        "200":
          description: OK
      security:
      - Foro Alura Auth: []
...
components:
  schemas:
    UsuarioPutRequest:
      required:
      - id
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        email:
          type: string
        contrasena:
          type: string
    UsuarioResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        email:
          type: string
        _links:
          type: object
          properties:
            self:
              $ref: '#/components/schemas/Link'
            usuarios:
              $ref: '#/components/schemas/Link'
...
    UsuarioPostRequest:
      required:
      - contrasena
      - email
      - nombre
      type: object
      properties:
        nombre:
          type: string
        email:
          type: string
        contrasena:
          type: string
...
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
    PagedUsuarioResponse:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            usuarios:
              type: array
              items:
                $ref: '#/components/schemas/UsuarioResponse'
        _links:
          type: object
          properties:
            self:
              $ref: '#/components/schemas/Link'
        page:
          $ref: '#/components/schemas/PageMetadata'
...
    Link:
      type: object
      properties:
        href:
          type: string
    ErrorResponse:
      type: object
      properties:
        campo: 
          type: string
        error:
          type: string
  securitySchemes:
    Foro Alura Auth:
      type: http
      description: "Ingresa el JWT Token. Este puede ser obtenido desde /login. Para\
        \ ejecutar pruebas utilizando la basede datos que se provee en el repositorio\
        \ [Foro Alura](https://github.com/SaulMMBP/foro-alura), puedes utilizar: \n\
        - Rol `ADMIN`\n```\n{\n\t'usuario':'saul@mail.com',\n\t'contrasena':'1234'\n\
        }\n```\n- Rol `INSTRUCTOR`\n```\n{\n\t'usuario':'daniel@mail.com',\n\t'contrasena':'1234'\n\
        }\n```\n- Rol `USER`\n```\n{\n\t'usuario':'will@mail.com',\n\t'contrasena':'1234'\n\
        }\n```\n"
      scheme: bearer
      bearerFormat: JWT